sources/pseudo_service/requirements.txt
django >= 1.11
djangorestframework
django-crispy-forms